Add files via upload
TIL about is how to choose list of li elements. and I can choose few ways of it and it will be li:first-child, li:last-child, li:nth-child() 
and sedondly TIL pseudo-class, there are 4 type of pseudo-class. Link, Visited, Hover, Active. LVHA is need always exact sequence.
each element is target on Link, Visited, Hover, Active. so if I use link, it will be a:link, and visited is about after you visited link. and Hover is when mous put on the link it will change, and active is if I click on link it will change

looks complicated at first. but once you get it. it's easy to understand haha 
keep going !