Add files via upload
TIL about priority in CSS. there are few types of selector, first one is ID selector. this is number one priority. and it will choose Last selector of ID selector if there is a lot of ID selector
Second is Class and pseudo-class, this is same like ID class. it will choose last selector if there are a lot of class and pseudo-classs
Third is element selector, and there are p, div,li,etc
last one is Universal selector but I haven't learn yet :D
and there are 2 more things, which is high priority then ID selector. that is inline style, and more high priority is declaration but this is only use in last step. normally we don't use it and yeah this is all I learn today